Reorganize to make auto_selfcal pip installable.
