+++ auto_selfcal/run_selfcal.py
@@ -1,14 +1,11 @@
 import numpy as np
-from scipy import stats
 import glob
 import sys
-#execfile('selfcal_helpers.py',globals())
-sys.path.append("./")
-from selfcal_helpers import *
-from gaincal_wrapper import gaincal_wrapper, generate_settings_for_combinespw_fallback
-from image_analysis_helpers import *
-from mosaic_helpers import *
-from applycal_wrapper import applycal_wrapper
+from .selfcal_helpers import *
+from .gaincal_wrapper import gaincal_wrapper, generate_settings_for_combinespw_fallback
+from .image_analysis_helpers import *
+from .mosaic_helpers import *
+from .applycal_wrapper import applycal_wrapper
 
 # Mac builds of CASA lack MPI and error without this try/except
 try:
