+++ pyproject.toml
@@ -1,0 +1,35 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "auto_selfcal"
+version = "3.0"
+authors = [
+  { name="John Tobin", email="jtobin@nrao.edu" },
+  { name="Patrick Sheehan", email="psheehan@nrao.edu" },
+]
+description = "Automated self-calibration of interferometric data"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+dependencies = [
+    "numpy",
+    "scipy",
+    "casatools",
+    "casatasks",
+    "casaplotms",
+    "casaviewer",
+    "protobuf<=3.20.3",
+]
+
+[project.urls]
+Homepage = "https://github.com/jjtobin/auto_selfcal"
+Issues = "https://github.com/jjtobin/auto_selfcal/issues"
+
+[project.scripts]
+auto_selfcal = "auto_selfcal.__main__:main_cli"
